adding search api functionality

+++ components/apimgt/org.wso2.carbon.apimgt.core/src/main/java/org/wso2/carbon/apimgt/core/impl/APIPublisherImpl.java
@@ -27,8 +27,10 @@ import org.wso2.carbon.apimgt.core.dao.ApiDAO;
 import org.wso2.carbon.apimgt.core.dao.ApplicationDAO;
 import org.wso2.carbon.apimgt.core.exception.APIManagementDAOException;
 import org.wso2.carbon.apimgt.core.exception.APIManagementException;
+import org.wso2.carbon.apimgt.core.exception.ApiDeleteFailureException;
 import org.wso2.carbon.apimgt.core.models.API;
 import org.wso2.carbon.apimgt.core.models.APIStatus;
+import org.wso2.carbon.apimgt.core.models.APISummaryResults;
 import org.wso2.carbon.apimgt.core.models.DocumentInfo;
 import org.wso2.carbon.apimgt.core.models.LifeCycleEvent;
 import org.wso2.carbon.apimgt.core.models.Provider;
@@ -36,6 +38,7 @@ import org.wso2.carbon.apimgt.core.models.Subscriber;
 import org.wso2.carbon.apimgt.core.util.APIUtils;
 
 import java.io.InputStream;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -340,25 +343,34 @@ public class APIPublisherImpl extends AbstractAPIManager implements APIPublisher
     @Override
     public void deleteAPI(String identifier) throws APIManagementException {
         try {
-            getApiDAO().deleteAPI(identifier);
-            APIUtils.logDebug("API with id " + identifier + " was deleted successfully.", log);
+            if (getAPISubscriptionCountByAPI(identifier) < 0) {
+                getApiDAO().deleteAPI(identifier);
+                APIUtils.logDebug("API with id " + identifier + " was deleted successfully.", log);
+            } else {
+                throw new ApiDeleteFailureException("API with " + identifier + " already have subscriptions");
+            }
         } catch (APIManagementDAOException e) {
             APIUtils.logAndThrowException("Error occurred while deleting the API with id " + identifier, e, log);
         }
     }
 
     /**
-     * Search API
-     *
-     * @param searchTerm Search Term
-     * @param searchType Search Type
-     * @param providerId
-     * @return Set of APIs
+     * @param limit
+     * @param offset
+     * @param query
+     * @return
      * @throws APIManagementException
      */
     @Override
-    public List<API> searchAPIs(String searchTerm, String searchType, String providerId) throws APIManagementException {
-        return null;
+    public APISummaryResults searchAPIs(Integer limit, Integer offset, String query) throws APIManagementException {
+
+        APISummaryResults apiSummaryResults = null;
+        try {
+            apiSummaryResults = getApiDAO().searchAPIs(APIUtils.createSearchCriteriaMap(query),offset,limit);
+        } catch (APIManagementDAOException e) {
+            APIUtils.logAndThrowException("Error occurred while Searching the API with query " + query, e, log);
+        }
+        return apiSummaryResults;
     }
 
     /**
